requirements.txt
langchain
python-dotenv
langchain_community
langchain-cohere
beautifulsoup4
faiss-cpu
openai
streamlit
langchain_openai
openpyxl
